Изменение : MP4Splitter (Bento4) - при расчете TimeStamp происходило переполнение переменной, поэтому вместо unsigned long long используем тип double.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@2430 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/ExtLib/Bento4/Core/Ap4Track.cpp
@@ -279,7 +279,11 @@ AP4_Track::GetSampleIndexForTimeStampMs(AP4_TimeStamp ts, AP4_Ordinal& index)
 AP4_Result  
 AP4_Track::GetSampleIndexForRefTime(REFERENCE_TIME rt, AP4_Ordinal& index)
 {
-    AP4_TimeStamp ts = (AP4_TimeStamp(rt) * m_MediaTimeScale + 5000000) / 10000000;
+    // MPC-BE custom code start
+    //AP4_TimeStamp ts = (AP4_TimeStamp(rt) * m_MediaTimeScale + 5000000) / 10000000;
+    AP4_TimeStamp ts = (AP4_TimeStamp)((double(rt) * m_MediaTimeScale + 5000000) / 10000000);
+    // need calculate in double, because the (AP4_TimeStamp(rt) * m_MediaTimeScale) can give overflow
+    // MPC-BE custom code end
     //AP4_TimeStamp ts = (AP4_TimeStamp)(double(rt) * m_MediaTimeScale / 10000000 + 0.5);
 
     return m_SampleTable->GetSampleIndexForTimeStamp(ts, index);
